src/main/python/requirements.txt
beautifulsoup4==4.7.1
chardet==3.0.4
decorator==4.4.0
idna==2.8
jieba==0.39
mysql-connector==2.1.6
networkx==2.2
requests==2.21.0
rwlock==0.0.7
soupsieve==1.8
threadpool==1.3.2
urllib3==1.24.1
